renaming the rpc server file
